# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorResponse {
  error: string;
}

export interface AccessPermissions {
  canManage: boolean;
}

export interface MentorOption {
  id: number;
  name: string;
}

export type ServerStatusStatus = typeof ServerStatusStatus[keyof typeof ServerStatusStatus];


export const ServerStatusStatus = {
  ok: 'ok',
} as const;

export interface ServerStatus {
  status: ServerStatusStatus;
}

export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  error: 'error',
} as const;

export type HealthStatusDatabase = typeof HealthStatusDatabase[keyof typeof HealthStatusDatabase];


export const HealthStatusDatabase = {
  connected: 'connected',
  disconnected: 'disconnected',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  database: HealthStatusDatabase;
}

/**
 * @nullable
 */
export type MentorMentorType = typeof MentorMentorType[keyof typeof MentorMentorType] | null;


export const MentorMentorType = {
  interno: 'interno',
  externo: 'externo',
} as const;

export interface Mentor {
  id: number;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  expertiseArea: string | null;
  /** @nullable */
  mentorType: MentorMentorType;
}

export interface Student {
  id: number;
  name: string;
  /** @minimum 0 */
  sortOrder: number;
}

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  pitchSummary: string | null;
  /** @nullable */
  currentStage: string | null;
  createdAt: string;
  mainMentor: Mentor;
  students: Student[];
  /** @minimum 0 */
  sessionCount: number;
}

export interface TeamInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @minimum 1 */
  mainMentorId: number;
  /**
     * @items.minLength 1
     * @items.maxLength 200
     */
  students: string[];
}

export interface TeamUpdateInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /** @minimum 1 */
  mainMentorId?: number;
  /**
     * @minLength 1
     * @maxLength 120
     */
  expectedName: string;
  /** @minimum 1 */
  expectedMainMentorId: number;
}

export interface StudentSnapshot {
  /** @minimum 1 */
  id: number;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @minimum 0 */
  sortOrder: number;
}

export interface TeamDeleteInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  expectedName: string;
  /** @minimum 1 */
  expectedMainMentorId: number;
  expectedStudents: StudentSnapshot[];
}

export interface StudentInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
}

export interface StudentUpdateInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  expectedName: string;
}

export interface StudentDeleteInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  expectedName: string;
}