# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: teams
    description: Mentoring teams
  - name: access
    description: User permissions
  - name: mentors
    description: Available mentors
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Server health check
      description: Returns whether the API process is available
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStatus"
  /health:
    get:
      operationId: getHealth
      tags: [health]
      summary: Check API and database connection
      responses:
        "200":
          description: API and database are connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Database is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /teams:
    get:
      operationId: getTeams
      tags: [teams]
      summary: List mentoring teams
      description: Returns each team with its lead mentor, ordered students, and session count.
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team with its initial students
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Invalid team data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "409":
          description: Team name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team name or its main mentor
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdateInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Invalid team data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "404":
          description: Team not found
        "409":
          description: Team name already exists or team changed since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team and its students
      description: Teams with recorded mentoring sessions cannot be deleted.
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamDeleteInput"
      responses:
        "204":
          description: Team deleted
        "400":
          description: Invalid deletion confirmation
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "404":
          description: Team not found
        "409":
          description: Team has mentoring sessions or changed since confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}/students:
    post:
      operationId: createStudent
      tags: [teams]
      summary: Add a student to a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Invalid student data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "404":
          description: Team not found
  /teams/{teamId}/students/{studentId}:
    patch:
      operationId: updateStudent
      tags: [teams]
      summary: Update a student's name
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdateInput"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Invalid student data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "404":
          description: Team or student not found
        "409":
          description: Student changed since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [teams]
      summary: Remove a student from a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentDeleteInput"
      responses:
        "204":
          description: Student deleted
        "400":
          description: Invalid deletion confirmation
        "401":
          description: Sign in required
        "403":
          description: Administrator access required
        "404":
          description: Team or student not found
        "409":
          description: Student changed since confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mentors:
    get:
      operationId: getMentors
      tags: [mentors]
      summary: List mentors available for team assignment
      responses:
        "200":
          description: Mentors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorOption"
        "401":
          description: Sign in required
        "403":
          description: Approved access required
  /access:
    get:
      operationId: getAccessPermissions
      tags: [access]
      summary: Get current user's safe access capabilities
      responses:
        "200":
          description: Access capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessPermissions"
        "401":
          description: Sign in required
        "403":
          description: Approved access required
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AccessPermissions:
      type: object
      properties:
        canManage:
          type: boolean
      required: [canManage]
    MentorOption:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    ServerStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
      required:
        - status
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
        database:
          type: string
          enum: [connected, disconnected]
      required:
        - status
        - database
    Mentor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        expertiseArea:
          type: ["string", "null"]
        mentorType:
          type: ["string", "null"]
          enum: [interno, externo, null]
      required:
        - id
        - name
        - email
        - expertiseArea
        - mentorType
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
          minimum: 0
      required:
        - id
        - name
        - sortOrder
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        pitchSummary:
          type: ["string", "null"]
        currentStage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        mainMentor:
          $ref: "#/components/schemas/Mentor"
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        sessionCount:
          type: integer
          minimum: 0
      required:
        - id
        - name
        - pitchSummary
        - currentStage
        - createdAt
        - mainMentor
        - students
        - sessionCount
    TeamInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        mainMentorId:
          type: integer
          minimum: 1
        students:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 200
      required: [name, mainMentorId, students]
    TeamUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        mainMentorId:
          type: integer
          minimum: 1
        expectedName:
          type: string
          minLength: 1
          maxLength: 120
        expectedMainMentorId:
          type: integer
          minimum: 1
      required: [expectedName, expectedMainMentorId]
    StudentSnapshot:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 200
        sortOrder:
          type: integer
          minimum: 0
      required: [id, name, sortOrder]
    TeamDeleteInput:
      type: object
      properties:
        expectedName:
          type: string
          minLength: 1
          maxLength: 120
        expectedMainMentorId:
          type: integer
          minimum: 1
        expectedStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentSnapshot"
      required: [expectedName, expectedMainMentorId, expectedStudents]
    StudentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
      required: [name]
    StudentUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        expectedName:
          type: string
          minLength: 1
          maxLength: 200
      required: [name, expectedName]
    StudentDeleteInput:
      type: object
      properties:
        expectedName:
          type: string
          minLength: 1
          maxLength: 200
      required: [expectedName]