# lib/api-zod/src/generated/types/studentDeleteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentDeleteInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  expectedName: string;
}