# lib/api-zod/src/generated/types/studentSnapshot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentSnapshot {
  /** @minimum 1 */
  id: number;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @minimum 0 */
  sortOrder: number;
}