# lib/api-zod/src/generated/types/teamDeleteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentSnapshot } from './studentSnapshot';

export interface TeamDeleteInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  expectedName: string;
  /** @minimum 1 */
  expectedMainMentorId: number;
  expectedStudents: StudentSnapshot[];
}